+++ gc_registry/account/routes.py
@@ -2,7 +2,8 @@ from esdbclient import EventStoreDBClient
 from fastapi import APIRouter, Depends, HTTPException
 from sqlmodel import Session
 
-from gc_registry.account import models
+from gc_registry.account.models import Account, AccountBase, AccountRead
+from gc_registry.account.schemas import AccountUpdate, AccountWhitelist
 from gc_registry.account.validation import (
     validate_account,
     validate_account_whitelist_update,
@@ -13,17 +14,15 @@ from gc_registry.core.database import db, events
 router = APIRouter(tags=["Accounts"])
 
 
-@router.post("/create", status_code=201, response_model=models.AccountRead)
+@router.post("/create", status_code=201, response_model=AccountRead)
 def create_account(
-    account_base: models.AccountBase,
+    account_base: AccountBase,
     write_session: Session = Depends(db.get_write_session),
     read_session: Session = Depends(db.get_read_session),
     esdb_client: EventStoreDBClient = Depends(events.get_esdb_client),
 ):
     validate_account(account_base, read_session)
-    accounts = models.Account.create(
-        account_base, write_session, read_session, esdb_client
-    )
+    accounts = Account.create(account_base, write_session, read_session, esdb_client)
     if not accounts:
         raise HTTPException(status_code=500, detail="Could not create Account")
 
@@ -32,26 +31,26 @@ def create_account(
     return account
 
 
-@router.get("/{account_id}", response_model=models.AccountRead)
+@router.get("/{account_id}", response_model=AccountRead)
 def read_account(
     account_id: int,
     read_session: Session = Depends(db.get_read_session),
 ):
-    account = models.Account.by_id(account_id, read_session)
+    account = Account.by_id(account_id, read_session)
     if not account:
         raise HTTPException(status_code=404, detail="Account not found")
     return account
 
 
-@router.patch("/update/{account_id}", response_model=models.AccountRead)
+@router.patch("/update/{account_id}", response_model=AccountRead)
 def update_account(
     account_id: int,
-    account_update: models.AccountUpdate,
+    account_update: AccountUpdate,
     write_session: Session = Depends(db.get_write_session),
     read_session: Session = Depends(db.get_read_session),
     esdb_client: EventStoreDBClient = Depends(events.get_esdb_client),
 ):
-    account = models.Account.by_id(account_id, write_session)
+    account = Account.by_id(account_id, write_session)
     if not account:
         raise HTTPException(
             status_code=404, detail=f"Account ID not found: {account_id}"
@@ -67,15 +66,15 @@ def update_account(
     return updated_account.model_dump()
 
 
-@router.patch("/update_whitelist/{account_id}", response_model=models.AccountRead)
+@router.patch("/update_whitelist/{account_id}", response_model=AccountRead)
 def update_whitelist(
     account_id: int,
-    account_whitelist_update: models.AccountWhitelist,
+    account_whitelist_update: AccountWhitelist,
     write_session: Session = Depends(db.get_write_session),
     read_session: Session = Depends(db.get_read_session),
     esdb_client: EventStoreDBClient = Depends(events.get_esdb_client),
 ):
-    account = models.Account.by_id(account_id, write_session)
+    account = Account.by_id(account_id, write_session)
     if not account:
         raise HTTPException(
             status_code=404, detail=f"Account ID not found: {account_id}"
@@ -85,7 +84,7 @@ def update_whitelist(
         account, account_whitelist_update, read_session
     )
 
-    account_update = models.AccountUpdate(account_whitelist=modified_whitelist)
+    account_update = AccountUpdate(account_whitelist=modified_whitelist)
 
     updated_account = account.update(
         account_update, write_session, read_session, esdb_client
@@ -97,9 +96,7 @@ def update_whitelist(
     return updated_account.model_dump()
 
 
-@router.delete(
-    "/delete/{account_id}", status_code=200, response_model=models.AccountRead
-)
+@router.delete("/delete/{account_id}", status_code=200, response_model=AccountRead)
 def delete_account(
     account_id: int,
     write_session: Session = Depends(db.get_write_session),
@@ -107,7 +104,7 @@ def delete_account(
     esdb_client: EventStoreDBClient = Depends(events.get_esdb_client),
 ):
     try:
-        account = models.Account.by_id(account_id, write_session)
+        account = Account.by_id(account_id, write_session)
         accounts = account.delete(write_session, read_session, esdb_client)
         if not accounts:
             raise ValueError(f"Account id {account_id} not found")
